+++ src/main/java/org/betonquest/betonquest/quest/condition/party/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * {@link org.betonquest.betonquest.api.quest.condition.online.OnlineCondition OnlineCondition}
+ * contains classes for party conditions.
+ */
+package org.betonquest.betonquest.quest.condition.party;