+++ CHANGELOG.md
@@ -20,6 +20,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - global variable support for npc ids and condition ids in `effectlib` section
 - `and` condition can now be static
 - missing tab-completion and documentation for globalpoints and globaltags command
+- `party` condition now supports a location variable, to be used in a static context
 ### Changed
 - `spawn` event now only spawn mobs and no other entities
 ### Deprecated